Create jar compatible with macos (#2)
* Replace Makefile to CMake for building jni
Create Mac-os build
Update ci to produce a jar containing both
Bump version to 3.0
Modify swig to handle C99 type (uint64, int64, ...)

* Include openmp libraries for linux

* Compile with O3

* Fix change

+++ src/test/java/com/vectorsearch/faiss/OneFlat.java
@@ -35,7 +35,7 @@ public class OneFlat extends FaissTestCase {
     public void train() {
         index.add(inputRowCount, inputVectors.cast());
         final boolean isTrained = index.getIs_trained();
-        final int nTotal = index.getNtotal();
+        final long nTotal = index.getNtotal();
         final String msg = "isTrained = " + isTrained + ", nTotal = " + nTotal;
         LOGGER.info(msg);
     }
